Classroom

+++ src/Homework/CalculatorV2.java
@@ -2,8 +2,8 @@ package Homework;
 
 public class CalculatorV2 {
 
-    public static float midArithmetic(int a, int b, int c){
-        return (a + b + c ) / 3f;
+    public static double midArithmetic(int a, int b, int c){
+        return (a + b + c ) / 3.0;
     }
     public static void main(String[] args) {
         System.out.println(midArithmetic(2,4,4));
